business-logic-container/requirements.txt
azure-iot-device>=2.12.0
shapely==1.8.2
numpy
